ai-service/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database
pymongo==4.6.0
redis==5.0.1

# Environment
python-dotenv==1.0.0

# ML & AI
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Audio Processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1

# HTTP Client
requests==2.31.0

# Utilities
python-dateutil==2.8.2
